+++ R/harmony_wrapper.R
@@ -27,12 +27,7 @@ harmony_init <- function(embedding, design_matrix,
   sigma <- rep_len(sigma, nclust)
   lambda_range = c(0.1, 10)
   if(packageVersion("harmony") < "1.2.0"){
-    harmonyObj <- harmony_new_object()
-    harmonyObj$setup(
-      embedding, phi,
-      sigma, theta, lambda_vec, max.iter.cluster, epsilon.cluster,
-      epsilon.harmony, nclust, block.size, lambda_range, B_vec, verbose
-    )
+    stop("Your 'harmony' version is outdated: ", packageVersion("harmony"), ". Please update to version >= 1.2.0")
   }else{
     alpha <- 0.2
     harmonyObj <- harmony::RunHarmony(embedding, mm_groups, nclust = nclust, max.iter = 0, return_object = TRUE, verbose = FALSE)
@@ -53,23 +48,14 @@ harmony_init <- function(embedding, design_matrix,
 #' @keywords internal
 harmony_new_object <- function(){
   Y <- randn(3, 100)
-  harmonyObj <- if(utils::packageVersion("harmony") >= "1.0.3"){
-    # Harmony ignores 'verbose = FALSE'
-    harmony::RunHarmony(Y, rep(c("a", "b"), length.out = 100), nclust = 2, max.iter = 0, return_object = TRUE, verbose = FALSE)
-  }else{
-    harmony::HarmonyMatrix(Y, rep(c("a", "b"), length.out = 100), do_pca = FALSE, nclust = 2, max.iter.harmony = 0, return_object = TRUE)
-  }
+  harmony::RunHarmony(Y, rep(c("a", "b"), length.out = 100), nclust = 2, max.iter = 0, return_object = TRUE, verbose = FALSE)
   harmonyObj
 }
 
 harmony_init_clustering <- function(harmonyObj, iter.max = 25, nstart = 10){
   stopifnot(is(harmonyObj, "Rcpp_harmony"))
   harmonyObj$Y <- t(stats::kmeans(t(harmonyObj$Z_cos), centers = harmonyObj$K, iter.max = iter.max, nstart = nstart)$centers)
-  if(utils::packageVersion("harmony") <= "1.1.0"){
-    harmonyObj$init_cluster_cpp(0)
-  }else{
-    harmonyObj$init_cluster_cpp()
-  }
+  harmonyObj$init_cluster_cpp()
   harmonyObj
 }