+++ NEWS.md
@@ -4,6 +4,8 @@
 the `group_by` argument for `find_de_neighborhoods`. (thanks Katha for pushing for this feature)
 * The formula parsing automatically detects global variables and adds them to the 
 colData. This avoids problems with the random test / training assignment.
+* Duplicate column names in colData are now longer allowed.
+
 
 # v1.1